+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-10T00:46:29.786Z",
+    "2026-02-10T09:05:30.137Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {