+++ packages/orq-rc/src/models/operations/fileget.ts
@@ -94,7 +94,7 @@ export const FileGetResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-10T00:46:33.266Z",
+    "2026-02-10T09:05:33.478Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {